iMaskVending22.READ.ME
Please know that this is an Alliance repository.
What that means for you is:
Any thing found in this repository has been created and placed here for the use of correcting the injustices being committed by those placed in positions to assist and help people; however, these people are doing the opposite by abusing their positions within the county human service level, helping themselves gain financially, leaving little or nothing for the ones they are supposed to be serving in the first place.  Greed is the root cause of this tragic practice and goes as far up the ladder that one could imagine.  Through forensic accounting practices, and ethically covert tactics both above board and below, realignment of injustices may be possible.  To catch a thief you have to think like one.  To catch a thief red handed, you have to be inside the vault they're about to open and steal from. Any contributions made to this project are greatly appreciated and is recognized as something good to support and stand with Alliance for what is right.  Thank You